Change: Move AsyncIteratorMock to pontos.testing module
Make the AsyncIteratorMock available for other applications then pontos.
